refactor: Organize folders and remove some warnings
