+++ app/repositories/user_repository.rb
