+++ app/services/authentication_service.rb
