+++ app/workers/create_product_worker.rb
