+++ docker-compose-test.yml
@@ -1,5 +1,3 @@
-version: "3.7"
-
 services:
   test-db:
     image: postgres:16
