+++ docker-compose.yml
@@ -1,5 +1,3 @@
-version: "3.7"
-
 services:
   redis:
     image: redis:alpine3.18
